Fixed resetting, this was an index problem.
Not sure about the rest, though. Having to put DlgOptionEnd() after every option is tiresome. Yet, it allows you to end the dialogue after an option, without returning to the main dialogue:

Example 1:
if (DlgOption)
{
	...;
	DlgOptionEnd();
}
DlgText(...); // this will be displayed, even after you choose the option

Example 2:
if (DlgOption)
{
	...; // The dialogue will end here
}
DlgText(...); // this will not be displayed, unless you chose an option with DlgOptionEnd()

Maybe a deliberate "do nothing" DlgStop() will help?

if (DlgOption)
{
	...; // The dialogue will continue with the text.
}
if (DlgOption)
{
	...;
	DlgStop(); // The dialogue will end here
}
DlgText(...);